# bin/main/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Decision Engine — Rule Platform API
  description: |
    # Decision Engine — Plataforma de Regras de PLD

    API da plataforma de decisão para Prevenção à Lavagem de Dinheiro (PLD).
    Desacopla a lógica de **detecção** (Screening Context) da lógica de **decisão**
    (quando gerar alertas, com quais critérios, baseado em quais fatos).

    ## Arquitetura

    ```
    ┌─────────────────┐     DetectionEvent      ┌──────────────────┐    DecisionMadeEvent    ┌───────────────┐
    │  Screening      │ ──────────────────────▶ │  Decision        │ ─────────────────────▶ │  Alert        │
    │  Context (MF09) │                         │  Context         │                        │  Context      │
    │                 │                         │                  │                        │               │
    │  Detecta        │                         │  Avalia regras   │                        │  Cria alertas │
    │  keywords       │                         │  Decide ações    │                        │  com status   │
    └─────────────────┘                         └──────────────────┘                        └───────────────┘
    ```

    ## Fluxo de Decisão

    ```
    1. Transação PIX recebida
    2. KeywordScreeningService detecta termo restrito → publica DetectionEvent
    3. Decision Engine consome evento:
       a) Verifica idempotência (transactionId + ruleId)
       b) Carrega configuração ativa da regra
...
    5. Publica DecisionMadeEvent
...

...

...

...

...
servers:
...
paths:
...
  /v1/decision/rules:
    get:
      tags: [Rule Catalog]
      summary: Listar definições de regras
      description: |
        Retorna todas as Rule Definitions cadastradas no sistema, opcionalmente filtradas
        por contexto (SCREENING, TRANSACTION, etc.) e/ou categoria (KEYWORD_SCREENING, AML, etc.).

        Rule Definitions são técnicas e somente leitura para analistas — criadas pela engenharia.
      operationId: listRules
      parameters:
        - name: context
          in: query
          required: false
          description: Filtrar por contexto da regra
          schema:
            $ref: '#/components/schemas/RuleContext'
          example: SCREENING
        - name: category
          in: query
          required: false
          description: Filtrar por categoria funcional
          schema:
            $ref: '#/components/schemas/RuleCategory'
          example: KEYWORD_SCREENING
      responses:
        '200':
          description: Lista de definições de regras (pode ser vazia)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleDefinitionResponse'
              examples:
                mvp:
                  summary: MVP — apenas KEYWORD_SCREENING
                  value:
                    - id: "550e8400-e29b-41d4-a716-446655440000"
                      code: "KEYWORD_SCREENING"
                      name: "Keyword Screening"
                      description: "Regra MF09 — detecção de termos restritos em descrições de transações PIX"
                      context: "SCREENING"
                      category: "KEYWORD_SCREENING"
                      supportedFacts: ["keywordMatched", "customerRisk"]
                      supportedActions: ["GENERATE_ALERT", "IGNORE"]
                      status: "ACTIVE"
                      createdAt: "2025-01-01T00:00:00Z"
  /v1/decision/rules/{code}:
    get:
      tags: [Rule Catalog]
      summary: Buscar definição de regra por código
      description: Retorna os detalhes de uma Rule Definition específica pelo seu código único.
      operationId: getRuleByCode
      parameters:
        - name: code
          in: path
          required: true
          description: "Código único da regra (ex.: KEYWORD_SCREENING)"
          schema:
            type: string
          example: KEYWORD_SCREENING
      responses:
        '200':
          description: Definição da regra encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleDefinitionResponse'
        '404':
          description: Regra não encontrada para o código informado

  # ═══════════════════════════════════════════════════════════════════════════════
  # FACT REGISTRY — Consulta de fatos e entidades
  # ═══════════════════════════════════════════════════════════════════════════════
...
  /v1/decision/rule-configurations/{id}/dry-run:
    post:
      tags: [Dry-Run]
      summary: Executar dry-run (simulação)
      description: |
        Avalia uma configuração com facts fornecidos manualmente pelo analista.
        Usa o mesmo motor de avaliação da produção (ExpressionEvaluator + RuleEngine),
        garantindo que o resultado do teste é idêntico ao comportamento real.

        ```
        ┌─────────────────────────────────────────────────────────────┐
        │ O que o Dry-Run FAZ:                                        │
        │ ✅ Avalia expressions com os mesmos algoritmos de produção   │
        │ ✅ Retorna decisão (ALERT/IGNORE) + detalhamento             │
        │ ✅ Persiste DryRunLog (para habilitar ativação)              │
        ├─────────────────────────────────────────────────────────────┤
        │ O que o Dry-Run NÃO FAZ:                                    │
        │ ❌ Não gera alertas                                          │
        │ ❌ Não persiste DecisionExecution                             │
        │ ❌ Não publica eventos                                        │
        │ ❌ Não invoca FactResolvers (facts são manuais)               │
        └─────────────────────────────────────────────────────────────┘
        ```

        Funciona para configurações **draft** e **active**.
      operationId: executeDryRun
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da configuração a ser testada
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DryRunRequest'
            examples:
              alertScenario:
                summary: Cenário que gera ALERT
                description: keyword match + risco alto (AR >= MR)
                value:
                  facts:
                    keywordMatched: true
                    customerRisk: "AR"
              ignoreScenario:
                summary: Cenário que gera IGNORE
                description: keyword match mas risco baixo (BR < MR)
                value:
                  facts:
                    keywordMatched: true
                    customerRisk: "BR"
              noKeyword:
                summary: Sem keyword match
                description: Mesmo com risco alto, sem keyword = IGNORE
                value:
                  facts:
                    keywordMatched: false
                    customerRisk: "AR"
      responses:
        '200':
          description: Resultado do dry-run com detalhamento completo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DryRunResponse'
              examples:
                alertResult:
                  summary: Resultado ALERT
                  value:
                    decision: "ALERT"
                    actions: ["GENERATE_ALERT"]
                    matchedExpressions:
                      - factName: "keywordMatched"
                        operator: "EQUALS"
                        expectedValue: true
                        actualValue: true
                        satisfied: true
                        justification: "Fact 'keywordMatched' (true) é igual a true"
                      - factName: "customerRisk"
                        operator: "GREATER_THAN_OR_EQUAL"
                        expectedValue: "MR"
                        actualValue: "AR"
                        satisfied: true
                        justification: "Fact 'customerRisk' (AR) é >= MR (ordinal 2 >= 1)"
                    failedExpressions: []
                    configurationVersion: 1
                ignoreResult:
                  summary: Resultado IGNORE (risco baixo)
                  value:
                    decision: "IGNORE"
                    actions: []
                    matchedExpressions:
                      - factName: "keywordMatched"
                        operator: "EQUALS"
                        expectedValue: true
                        actualValue: true
                        satisfied: true
                        justification: "Fact 'keywordMatched' (true) é igual a true"
                    failedExpressions:
                      - factName: "customerRisk"
                        operator: "GREATER_THAN_OR_EQUAL"
                        expectedValue: "MR"
                        actualValue: "BR"
                        satisfied: false
                        justification: "Fact 'customerRisk' (BR) não é >= MR (ordinal 0 < 1)"
                    configurationVersion: 1
        '404':
          description: Configuração não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Facts inválidos. Possíveis causas:
            - Fact name não existe no catálogo
            - Fact está desabilitado
            - Tipo do valor incompatível (ex.: enviar String para fact BOOLEAN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════════
  # DECISION EXECUTION — Auditoria
  # ═══════════════════════════════════════════════════════════════════════════════
...
components:
  schemas:
...
    RuleContext:
      type: string
      description: |
        Contexto ao qual uma Rule Definition pertence.
        Identifica o domínio de avaliação da regra.
      enum: [SCREENING, TRANSACTION, CUSTOMER, ACCOUNT]
      example: SCREENING
    RuleCategory:
      type: string
      description: |
        Categoria funcional de uma regra.
        Facilita filtros na UI e agrupamento lógico.
      enum: [KEYWORD_SCREENING, SANCTIONS, AML, FRAUD, VELOCITY]
      example: KEYWORD_SCREENING
    Decision:
      type: string
      description: |
        Resultado da avaliação do Decision Engine.
        - **ALERT**: Todas as condições satisfeitas → ação executada
        - **IGNORE**: Pelo menos uma condição falhou → sem ação
        - **REVIEW**: (futuro) Enviar para revisão manual
        - **BLOCK**: (futuro) Bloquear transação
      enum: [ALERT, IGNORE, REVIEW, BLOCK]
      example: ALERT
...
    ComparisonOperator:
      type: string
      description: |
        Operador de comparação para expressões.
        **MVP suporta:** EQUALS, NOT_EQUALS, GREATER_THAN_OR_EQUAL.
        Outros operadores estão modelados para evolução futura.
      enum: [EQUALS, NOT_EQUALS, GREATER_THAN, GREATER_THAN_OR_EQUAL, LESS_THAN, LESS_THAN_OR_EQUAL, IN, NOT_IN, CONTAINS]
      example: EQUALS
    Action:
      type: string
      description: |
        Ação a ser executada quando a decisão é ALERT.
        - **GENERATE_ALERT**: Cria alerta no Alert Context
        - **IGNORE**: Nenhuma ação (apenas registro)
        - **REVIEW**: (futuro) Enviar para revisão
        - **BLOCK**: (futuro) Bloquear
      enum: [GENERATE_ALERT, IGNORE, REVIEW, BLOCK]
      example: GENERATE_ALERT
...
    DryRunRequest:
      type: object
      description: |
        Request para executar dry-run com facts manuais.
        Os facts simulam o contexto de uma transação real.
        Cada fact deve ter nome e valor compatível com o Fact Registry.
      required: [facts]
      properties:
        facts:
          type: object
          description: |
            Mapa de facts (nome → valor) fornecidos manualmente.
            Tipos aceitos por fact:
            - keywordMatched: Boolean (true/false)
            - customerRisk: String enum ("BR", "MR", "AR")
          minProperties: 1
          additionalProperties: true
          example:
            keywordMatched: true
            customerRisk: "MR"
...
    RuleDefinitionResponse:
      type: object
      description: |
        Definição técnica de uma regra. Imutável para analistas.
        Criada e mantida pela engenharia.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único da regra
        code:
          type: string
          description: "Código legível da regra (ex.: KEYWORD_SCREENING). Único no sistema."
          example: "KEYWORD_SCREENING"
        name:
          type: string
          description: Nome descritivo da regra
          example: "Keyword Screening"
        description:
          type: string
          description: Descrição detalhada do propósito da regra
          example: "Regra MF09 — detecção de termos restritos em transações PIX"
        context:
          $ref: '#/components/schemas/RuleContext'
          description: Contexto de avaliação (SCREENING, TRANSACTION, etc.)
        category:
          $ref: '#/components/schemas/RuleCategory'
          description: Categoria funcional para agrupamento
        supportedFacts:
          type: array
          description: |
            Facts que podem ser usados em configurações desta regra.
            O analista só pode usar estes facts ao criar expressões.
          items:
            type: string
          example: ["keywordMatched", "customerRisk"]
        supportedActions:
          type: array
          description: Ações que esta regra suporta como resultado
          items:
            $ref: '#/components/schemas/Action'
          example: ["GENERATE_ALERT", "IGNORE"]
        status:
          type: string
          description: |
            Status da regra:
            - ACTIVE: disponível para configuração
            - INACTIVE: temporariamente desabilitada
            - DEPRECATED: não deve ser usada em novas configs
          enum: [ACTIVE, INACTIVE, DEPRECATED]
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
          description: Data de criação da definição
          example: "2025-01-01T00:00:00Z"
...
    DryRunResponse:
      type: object
      description: |
        Resultado completo do dry-run mostrando exatamente o que aconteceria em produção
        com os facts fornecidos. Inclui detalhamento de cada expressão avaliada.
      properties:
        decision:
          $ref: '#/components/schemas/Decision'
          description: |
            Decisão que seria tomada em produção:
            - ALERT: todas condições satisfeitas
            - IGNORE: pelo menos uma falhou
        actions:
          type: array
          description: Ações que seriam executadas (vazio se IGNORE)
          items:
            $ref: '#/components/schemas/Action'
        matchedExpressions:
          type: array
          description: Expressões que foram satisfeitas (com justificativa)
          items:
            $ref: '#/components/schemas/ExpressionEvaluationResponse'
        failedExpressions:
          type: array
          description: Expressões que falharam (com motivo da falha)
          items:
            $ref: '#/components/schemas/ExpressionEvaluationResponse'
        configurationVersion:
          type: integer
          description: Versão da configuração que foi testada
          example: 1
    ExpressionEvaluationResponse:
      type: object
      description: |
        Resultado da avaliação de uma expressão individual.
        Mostra o valor esperado vs. real e a justificativa em linguagem humana.
      properties:
        factName:
          type: string
          description: Nome do fact avaliado
          example: "customerRisk"
        operator:
          $ref: '#/components/schemas/ComparisonOperator'
          description: Operador utilizado na comparação
        expectedValue:
          description: Valor esperado pela expressão (conforme configurado)
          example: "MR"
        actualValue:
          description: Valor real encontrado no contexto (null se fact ausente)
          example: "AR"
          nullable: true
        satisfied:
          type: boolean
          description: Se a condição foi satisfeita (true) ou não (false)
          example: true
        justification:
          type: string
          description: |
            Explicação em linguagem humana do resultado da avaliação.
            Ex.: "Fact 'customerRisk' (AR) é >= MR (ordinal 2 >= 1)"
          example: "Fact 'customerRisk' (AR) é >= MR (ordinal 2 >= 1)"
...
    ErrorResponse:
      type: object
      description: |
        Formato padronizado de erro para todos os endpoints.
        Inclui timestamp, status HTTP, código de erro e mensagem descritiva.
      properties:
        timestamp:
          type: string
          format: date-time
          description: Momento do erro (ISO-8601)
          example: "2025-01-15T10:30:00Z"
        status:
          type: integer
          description: Código HTTP do erro
          example: 422
        error:
          type: string
          description: Descrição curta do tipo de erro HTTP
          example: "Unprocessable Entity"
        message:
          type: string
          description: Mensagem detalhada do erro (legível)
          example: "Fato 'unknownFact' não encontrado no catálogo"
        details:
          type: string
          description: "Detalhes adicionais (ex.: campos com erro de validação)"
          nullable: true
          example: "expressions[0].factName: não pode ser vazio; createdBy: é obrigatório"